% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{efetch}
\alias{efetch}
\title{Fetch records from Entrez}
\usage{
efetch(
  id_set,
  ...,
  retstart = 0L,
  retmax = 500L,
  retmode = "xml",
  .method = NA,
  .cookies = NA,
  .paginate = TRUE,
  .process = "identity",
  .call = rlang::caller_env()
)
}
\arguments{
\item{id_set}{ID set object.}

\item{...}{additional API parameters (refer to Entrez documentation).
Any set to \code{NULL} are removed.}

\item{retstart}{integer: index of first result (starts from 0).}

\item{retmax}{integer: maximum number of results to return.}

\item{retmode}{character: currently only \code{"xml"} is supported.}

\item{.method}{HTTP verb. If \code{NA}, a sensible default is chosen based on the request parameters.}

\item{.cookies}{path to persist cookies.
If \code{NULL}, cookies are not added to the request.
For helper functions: when \code{NA}, a temporary file is created (in this case only,
the temporary file will be cleaned up once all requests are performed).}

\item{.paginate}{logical: when \code{TRUE} send as many queries as needed to get the complete result.}

\item{.process}{function that processes the API results.
Can be a function or builtin processor as described in \code{\link{process}}.}

\item{.call}{call environment to use in error messages/traces.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}
}
\value{
output of \code{.process} from each page of results, combined with \code{\link[vctrs:vec_chop]{vctrs::list_unchop()}}.
}
\description{
Fetching can be slow, and Entrez will time out requests that take too long.
This helper supports pagination if you specify \code{retmax}.
}
\seealso{
Other API methods: 
\code{\link{einfo}()},
\code{\link{elink}()},
\code{\link{entrez_count}()},
\code{\link{entrez_translate}()},
\code{\link{epost}()},
\code{\link{esearch}()},
\code{\link{esummary}()}
}
\concept{API methods}
