% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{einfo}
\alias{einfo}
\alias{einfo_databases}
\title{Get details about Entrez databases}
\usage{
einfo(db, ..., retmode = "xml", version = "2.0", .call = rlang::current_env())

einfo_databases(..., retmode = "xml", .call = rlang::current_env())
}
\arguments{
\item{db}{name of database to provide information about.}

\item{...}{additional API parameters (refer to Entrez documentation).
Any set to \code{NULL} are removed.}

\item{retmode}{response format.}

\item{version}{response format version.}

\item{.call}{call environment to use in error messages/traces.
See \link[rlang:topic-error-call]{rlang::topic-error-call} and the \code{call} argument of \code{\link[cli:cli_abort]{cli::cli_abort()}}.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}
}
\value{
Character vector of database names for \code{einfo_databases()}.
An XML document with root node \verb{<eInfoResult>} for \code{einfo()}.
}
\description{
These functions call the EInfo endpoint. \code{einfo()} provides the number
of entries in the databases, the name and description, list of terms
usable in the query syntax, and list of link names usable with the ELink
endpoint.
}
\seealso{
Other API methods: 
\code{\link{efetch}()},
\code{\link{elink}()},
\code{\link{entrez_validate}()},
\code{\link{epost}()},
\code{\link{esearch}()},
\code{\link{esummary}()}
}
\concept{API methods}
