% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{entrez_count}
\alias{entrez_count}
\title{Count entries on the history server without returning them}
\usage{
entrez_count(id_set, .call = rlang::caller_env())
}
\arguments{
\item{id_set}{an \code{entrez_web_history} object.}

\item{.call}{call environment to use in error messages/traces.
See \link[rlang:topic-error-call]{rlang::topic-error-call} and the \code{call} argument of \code{\link[cli:cli_abort]{cli::cli_abort()}}.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}
}
\value{
integer number of entries.
}
\description{
Count entries on the history server without returning them
}
\seealso{
Other API methods: 
\code{\link{efetch}()},
\code{\link{einfo}()},
\code{\link{elink}()},
\code{\link{entrez_translate}()},
\code{\link{epost}()},
\code{\link{esearch}()},
\code{\link{esummary}()}
}
\concept{API methods}
