% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_set.R
\name{entrez_id_list}
\alias{entrez_id_list}
\alias{entrez_web_history}
\alias{is.entrez_id_set}
\alias{is.entrez_id_list}
\alias{is.entrez_web_history}
\alias{entrez_ids}
\alias{entrez_database}
\title{Entrez identifier lists}
\usage{
entrez_id_list(db, ids)

entrez_web_history(db, query_key, WebEnv, length = NA)

is.entrez_id_set(x)

is.entrez_id_list(x)

is.entrez_web_history(x)

entrez_ids(id_set)

entrez_database(id_set)
}
\arguments{
\item{db}{name of the associated Entrez database (e.g. \code{"biosample"}).}

\item{ids}{UIDs, coercible to a character vector (can be accessions or GI numbers).}

\item{query_key, WebEnv}{history server tokens returned by another Entrez API call}

\item{length}{number of UIDs in the set, if known}

\item{x}{object to test.}

\item{id_set}{ID set object.}
}
\description{
Many Entrez APIs accept either a UID list or tokens that point to a result stored
on its history server. The S3 classes here wrap these and keep track of the
database name that the identifiers belong to. Most of the API helpers in this
package are generic over the type of ID set and so can be used the same way with
either type. For large numbers of UIDs, the web history versions are generally
recommended to avoid running into server-side request timeouts.
}
\details{
\code{entrez_id_list} can be sliced and indexed like a vector (e.g. \code{id_set[1:10]} or
\code{tail(id_set)}).

\code{entrez_web_history} is an opaque reference to an object stored on the Entrez
history server. Unless the length is known at the time the object is
constructed, it will be missing.

Convert \code{entrez_id_list} to \code{entrez_web_history} with \code{\link[=epost]{epost()}}.

Convert \code{entrez_web_history} to \code{entrez_id_list} with \code{\link[=entrez_translate]{entrez_translate()}}.
}
