% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{entrez_request}
\alias{entrez_request}
\alias{entrez_api_key}
\title{Construct a request to the Entrez API}
\usage{
entrez_request(
  endpoint,
  ...,
  .method = "GET",
  .multi = "comma",
  .cookies = NULL,
  .verbose = getOption("jentre.verbose", default = TRUE),
  .call = current_env()
)

entrez_api_key(default = NULL)
}
\arguments{
\item{endpoint}{Entrez endpoint name (e.g. \code{"efetch.fcgi"}).}

\item{...}{additional API parameters (refer to Entrez documentation).
Any set to \code{NULL} are removed.}

\item{.method}{HTTP verb.
For \code{"POST"}, any params with vector values (usually just \code{id}) are sent in the
request body instead of the URL.}

\item{.multi}{controls how repeated params are handled (see \code{\link[httr2:req_url]{httr2::req_url_query()}}).}

\item{.cookies}{path to persist cookies.
If \code{NULL}, cookies are not added to the request.
For helper functions: when \code{NA}, a temporary file is created (in this case only,
the temporary file will be cleaned up once all requests are performed).}

\item{.verbose}{logical: when TRUE logs all API requests as messages in a compact format.
This uses a summarised format that does not include the request body for POST.
Use normal httr verbosity controls (e.g. \code{\link[httr2:with_verbosity]{httr2::local_verbosity()}}) to override
this behaviour and see more details.}

\item{.call}{call environment to use in error messages/traces.
See \link[rlang:topic-error-call]{rlang::topic-error-call} and the \code{call} argument of \code{\link[cli:cli_abort]{cli::cli_abort()}}.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}

\item{default}{default value to return if no global configuration is found.}
}
\value{
\itemize{
\item for \code{entrez_request()} an \code{httr2::request} object.
\item for \code{entrez_api_key()}, the API key as a character, or \code{default} if no global config exists.
}
}
\description{
This is a low-level helper that builds a request object but does not
perform the request. In general you'll likely use higher-level methods
like \code{\link[=efetch]{efetch()}} instead.
}
\details{
\code{email}, \code{tool}, and \code{api_key} have default values but these can be
overridden, or can be removed by setting them to \code{NULL}.
}
\section{API limits}{

The Entrez APIs are rate limited.
Requests in this package respect the API headers returned by Entrez.
Without an API key you will be rate limited more aggressively, so it is
recommended to \href{https://support.nlm.nih.gov/kbArticle/?pn=KA-05317}{obtain an API key}.
\code{jentre} searches for the API key in the following order:
\itemize{
\item the API parameter \code{entrez_key} provided to any API request function,
\item the \link[base:options]{option} \code{"jentre.api_key"}, then
\item the environment variable \code{ENTREZ_KEY}.
}

You can check the value is found properly using \code{entrez_api_key()}.
If no API key is set, a warning will be displayed. This can be suppressed
by setting the option \code{"jentre.silence_api_warning"} to \code{TRUE}.
}

