% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{entrez_translate}
\alias{entrez_translate}
\title{Convert web history result or accessions and other IDs to Entrez UIDs}
\usage{
entrez_translate(
  id_set,
  .paginate = 5000L,
  .path = NULL,
  .call = rlang::current_env()
)
}
\arguments{
\item{id_set}{an \code{\link{entrez_id_list}} object.}

\item{.paginate}{controls how multiple API requests are used to complete the call.
Pagination is performed using the \code{retstart} and \code{retmax} API parameters.
When set to an integer, no more than \code{.pagniate} items will be requested per API call.
When \code{FALSE} or \code{0}, only one API request is sent.}

\item{.path}{path specification for saving raw responses.
See \code{path} argument of \code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}}.}

\item{.call}{call environment to use in error messages/traces.
See \link[rlang:topic-error-call]{rlang::topic-error-call} and the \code{call} argument of \code{\link[cli:cli_abort]{cli::cli_abort()}}.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}
}
\description{
Convert web history result or accessions and other IDs to Entrez UIDs
}
\seealso{
Other API methods: 
\code{\link{efetch}()},
\code{\link{einfo}()},
\code{\link{elink}()},
\code{\link{entrez_count}()},
\code{\link{epost}()},
\code{\link{esearch}()},
\code{\link{esummary}()}
}
\concept{API methods}
