% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{epost}
\alias{epost}
\title{Register UIDs with the Entrez history server}
\usage{
epost(id_set, ..., WebEnv = NULL, .call = rlang::current_env())
}
\arguments{
\item{id_set}{an \code{\link{entrez_id_list}} object.}

\item{...}{additional API parameters (refer to Entrez documentation).
Any set to \code{NULL} are removed.}

\item{WebEnv}{either a character to pass on as-is, or an \code{\link{entrez_web_history}} object.}

\item{.call}{call environment to use in error messages/traces.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}
}
\description{
Register UIDs with the Entrez history server
}
\seealso{
Other API methods: 
\code{\link{efetch}()},
\code{\link{einfo}()},
\code{\link{elink}()},
\code{\link{entrez_count}()},
\code{\link{entrez_translate}()},
\code{\link{esearch}()},
\code{\link{esummary}()}
}
\concept{API methods}
