% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{esearch}
\alias{esearch}
\title{Search Entrez databases}
\usage{
esearch(
  term,
  db,
  ...,
  retstart = 0L,
  retmax = 1000L,
  retmode = "xml",
  usehistory = TRUE,
  WebEnv = NULL,
  query_key = NULL,
  .cookies = NA,
  .paginate = TRUE,
  .call = rlang::caller_env()
)
}
\arguments{
\item{term}{search query.}

\item{db}{Entrez database name.}

\item{...}{additional API parameters (refer to Entrez documentation).
Any set to \code{NULL} are removed.}

\item{retstart}{integer: index of first result (starts from 0).}

\item{retmax}{integer: maximum number of results to return.
Ignored when \code{usehistory} is \code{TRUE}.}

\item{retmode}{character: currently only \code{"xml"} is supported.}

\item{usehistory}{logical: when \code{TRUE} use the history server to return the result.}

\item{WebEnv, query_key}{either characters to pass on as-is, or \code{\link{entrez_web_history}} objects.}

\item{.cookies}{path to persist cookies.
If \code{NULL}, cookies are not added to the request.
For helper functions: when \code{NA}, a temporary file is created (in this case only,
the temporary file will be cleaned up once all requests are performed).}

\item{.paginate}{logical: when \code{TRUE} send as many queries as needed to get the complete result.
Ignored when \code{usehistory} is \code{TRUE}.}

\item{.call}{call environment to use in error messages/traces.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}
}
\value{
id set object (either a web history token or explicit list).
}
\description{
The search term field names are documented in the EInfo API endpoint:
see \code{\link[=einfo]{einfo()}}.
}
\seealso{
\url{https://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.ESearch}

Other API methods: 
\code{\link{efetch}()},
\code{\link{einfo}()},
\code{\link{elink}()},
\code{\link{entrez_count}()},
\code{\link{entrez_translate}()},
\code{\link{epost}()},
\code{\link{esummary}()}
}
\concept{API methods}
