% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_set.R, R/convert.R
\name{id_list}
\alias{id_list}
\alias{web_history}
\alias{is_id_set}
\alias{is_id_list}
\alias{is_web_history}
\alias{as_id_list}
\title{Entrez identifier sets}
\usage{
id_list(db, ids = character())

web_history(db, WebEnv, query_key, length = NA)

is_id_set(x)

is_id_list(x)

is_web_history(x)

as_id_list(x, .paginate = 5000L, .path = NULL, .call = current_env())
}
\arguments{
\item{db}{name of the associated Entrez database (e.g. \code{"biosample"}).}

\item{ids}{UIDs, coercible to a character vector (can be accessions or GI numbers).}

\item{query_key, WebEnv}{history server tokens returned by another Entrez API call.}

\item{length}{number of UIDs in the set, if known.}

\item{x}{object to test or convert.}

\item{.paginate}{controls how multiple API requests are used to complete the call.
Pagination is performed using the \code{retstart} and \code{retmax} API parameters.
When set to an integer, no more than \code{.pagniate} items will be requested per API call.
When \code{FALSE} or \code{0}, only one API request is sent.
Ignored when \code{usehistory} is \code{TRUE}.}

\item{.path}{path specification for saving raw responses.
See \code{path} argument of \code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}}.}

\item{.call}{call environment to use in error messages/traces.
See \link[rlang:topic-error-call]{rlang::topic-error-call} and the \code{call} argument of \code{\link[cli:cli_abort]{cli::cli_abort()}}.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}
}
\value{
\itemize{
\item For \code{id_list()} and \code{as_id_list()} an \code{id_list} vector.
\item For \code{web_history()} a \code{web_history} object.
\item For \code{is_id_set()}, \code{is_id_list()}, and \code{is_web_history()} a logical.
}
}
\description{
Many Entrez APIs accept either a UID list or tokens that point to a result stored
on its history server. The classes here wrap these and keep track of the
database name that the identifiers belong to. Most of the API helpers in this
package are generic over the type of ID set and so can be used the same way with
either type. For large numbers of UIDs, the web history versions are generally
recommended to avoid running into server-side request timeouts.
}
\details{
\code{id_list} is a vector and can be manipulated to take subsets (e.g. \code{id_set[1:10]} or
\code{tail(id_set)}).

\code{web_history} is an opaque reference to an ID list stored on the Entrez
history server. Through the course of API calls, information about the length or
the actual list of IDs may be discovered and cached, avoiding subsequent API calls.
\code{as_id_list()} can be used to extract the list of IDs.

Convert \code{id_list} to \code{web_history} with \code{\link[=epost]{epost()}}.
Convert \code{web_history} to \code{id_list} with \code{as_id_list()}.
}
\seealso{
\code{\link[=entrez_translate]{entrez_translate()}} and \code{\link[=entrez_count]{entrez_count()}}
}
