% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jentre-package.R
\docType{package}
\name{jentre-package}
\alias{jentre}
\alias{jentre-package}
\title{jentre: Toolkit for the 'Entrez' API}
\description{
A client for interacting with NCBI's 'Entrez' API \url{https://www.ncbi.nlm.nih.gov/books/NBK25499/}.
}
\author{
\strong{Maintainer}: Carl Suster \email{Carl.Suster@health.nsw.gov.au} (\href{https://orcid.org/0000-0001-7021-9380}{ORCID})

Other contributors:
\itemize{
  \item Western Sydney Local Health District, NSW Health [copyright holder]
}

}
\keyword{internal}
