% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_identity}
\alias{process_identity}
\title{Process API results}
\usage{
process_identity(doc)
}
\arguments{
\item{doc}{any parsed Entrez response document (e.g. {xml2} document or string)}
}
\value{
List of size 1 containing \code{doc}
}
\description{
Function to turn the parsed response document into meaningful data.
It must accept one argument, \code{doc}, the parsed response document.
The return value must be compatible with \code{\link[vctrs:vec_chop]{vctrs::list_unchop()}},
e.g. a vector, list, or data frame.
}
\details{
API results are parsed based on the \code{retmode} parameter. XML documents will
be parsed into \code{xml2::xml_document} objects and an error will be raised if
it contains an \verb{<ERROR>} node.

Builtin processors can be referred to by name instead of specifying your own
function. Some helpers provide additional processors, but these are always
available:
\itemize{
\item \code{"identity"}:
Puts the parsed output document into a list. Where multiple requests are made
(e.g. using the batched APIs like \code{\link[=efetch]{efetch()}}) these will then be
concatenated into a single list.
}
}
