% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_set.R
\name{check_id_set}
\alias{check_id_set}
\title{Check ID set is well formed}
\usage{
check_id_set(id_set, database = NULL, .call = rlang::caller_env())
}
\arguments{
\item{id_set}{ID set object.}

\item{database}{name of intended database.
If \code{NULL} the database name is not checked.}
}
\value{
\code{id_set}. This function raises an error if any check fails.
}
\description{
Check ID set is well formed
}
