% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{efetch}
\alias{efetch}
\title{Fetch records from Entrez}
\usage{
efetch(
  id_set,
  ...,
  retstart = 0L,
  retmax = NA,
  retmode = "xml",
  rettype = NULL,
  .method = NA,
  .cookies = NA,
  .paginate = 200L,
  .process = NA,
  .progress = "Fetching",
  .path = NULL,
  .call = rlang::current_env()
)
}
\arguments{
\item{id_set}{ID set object.}

\item{...}{additional API parameters (refer to Entrez documentation).
Any set to \code{NULL} are removed.}

\item{retstart}{integer: index of first result (starts from 0).}

\item{retmax}{integer: maximum number of results to return.
When \code{NA} this returns all results. When \code{NULL}, uses the Entrez default (typically 20).
Note that when using pagination with web history, it is possible that slightly more than
\code{retmax} results will be returned.}

\item{retmode}{character: requested document file format.}

\item{rettype}{character: requested document type.}

\item{.method}{HTTP verb. If \code{NA}, a sensible default is chosen based on the request parameters.}

\item{.cookies}{path to persist cookies.
If \code{NULL}, cookies are not added to the request.
For helper functions: when \code{NA}, a temporary file is created (in this case only,
the temporary file will be cleaned up once all requests are performed).}

\item{.paginate}{controls how multiple API requests are used to complete the call.
Pagination is performed using the \code{retstart} and \code{retmax} API parameters.
When set to an integer, no more than \code{.pagniate} items will be requested per API call.
When \code{FALSE} or \code{0}, only one API request is sent.}

\item{.process}{function that processes the API results.
Can be a function or builtin processor as described in \code{\link{process}}.
Additional builtin processors are available:
\itemize{
\item \code{"uilist"} to extract a list of IDs (suitable for \code{rettype = "uilist"}),
\item \code{NA} to use a sensible choice based on parameters. In particular, for \code{"uilist"}
requests, it will return an \code{\link{id_list}} object.
}}

\item{.progress}{controls progress bar; see the \code{progress} argument of
\code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}}.}

\item{.path}{path specification for saving raw responses.
See \code{path} argument of \code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}}.}

\item{.call}{call environment to use in error messages/traces.
See \link[rlang:topic-error-call]{rlang::topic-error-call} and the \code{call} argument of \code{\link[cli:cli_abort]{cli::cli_abort()}}.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}
}
\value{
output of \code{.process} from each page of results, combined with \code{\link[vctrs:vec_chop]{vctrs::list_unchop()}}.
}
\description{
Fetching can be slow, and Entrez will time out requests that take too long.
This helper supports pagination if you specify \code{retmax}.
}
\seealso{
Other API methods: 
\code{\link{einfo}()},
\code{\link{elink}()},
\code{\link{entrez_validate}()},
\code{\link{epost}()},
\code{\link{esearch}()},
\code{\link{esummary}()}
}
\concept{API methods}
