% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{entrez_count}
\alias{entrez_count}
\title{Count the number of entries in an ID set}
\usage{
entrez_count(id_set, .call = rlang::caller_env())
}
\arguments{
\item{id_set}{an \code{\link{id_set}} object.}

\item{.call}{call environment to use in error messages/traces.
See \link[rlang:topic-error-call]{rlang::topic-error-call} and the \code{call} argument of \code{\link[cli:cli_abort]{cli::cli_abort()}}.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}
}
\value{
integer number of entries.
}
\description{
If \code{id_set} is an \code{id_list} then this is equivalent to \code{length()}.
If it is a \code{web_history}, this may involve an Entrez API call to get the
number of entries. In this case the result is cached so that subsequent
calls don't hit the API again.
}
