% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{entrez_translate}
\alias{entrez_translate}
\title{Convert web history result or accessions and other IDs to Entrez UIDs}
\usage{
entrez_translate(id_set, retmax = 10000L, .call = rlang::caller_env())
}
\arguments{
\item{id_set}{ID set object.}

\item{retmax}{integer: maximum number of results to return.}

\item{.call}{call environment to use in error messages/traces.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}
}
\description{
Convert web history result or accessions and other IDs to Entrez UIDs
}
\seealso{
Other API methods: 
\code{\link{efetch}()},
\code{\link{einfo}()},
\code{\link{elink}()},
\code{\link{entrez_count}()},
\code{\link{epost}()},
\code{\link{esearch}()},
\code{\link{esummary}()}
}
\concept{API methods}
