% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{entrez_translate}
\alias{entrez_translate}
\title{Look up accessions and other IDs on Entrez}
\usage{
entrez_translate(
  id_set,
  idtype = NULL,
  .paginate = 5000L,
  .path = NULL,
  .call = rlang::current_env()
)
}
\arguments{
\item{id_set}{an \code{\link{entrez_id_list}} object.}

\item{idtype}{if \code{NULL} return Entrez numeric IDs; if \code{"acc"} return accessions.}

\item{.paginate}{controls how multiple API requests are used to complete the call.
Pagination is performed using the \code{retstart} and \code{retmax} API parameters.
When set to an integer, no more than \code{.pagniate} items will be requested per API call.
When \code{FALSE} or \code{0}, only one API request is sent.}

\item{.path}{path specification for saving raw responses.}

\item{.call}{call environment to use in error messages/traces.
See \link[rlang:topic-error-call]{rlang::topic-error-call} and the \code{call} argument of \code{\link[cli:cli_abort]{cli::cli_abort()}}.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}
}
\value{
\code{\link{id_list}} object
}
\description{
Passes the provided IDs through Entrez which has the effect of normalising the
accepted UIDs to the requested \code{idtype}, and rejecting invalid UIDs.
For web history lists, this forces results to be freshly downloaded
(unlike \code{\link[=as_id_list]{as_id_list()}} which can use cached results).
}
\seealso{
Other API methods: 
\code{\link{efetch}()},
\code{\link{einfo}()},
\code{\link{elink}()},
\code{\link{epost}()},
\code{\link{esearch}()},
\code{\link{esummary}()}
}
\concept{API methods}
