% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{esearch}
\alias{esearch}
\title{Search Entrez databases}
\usage{
esearch(
  term,
  db,
  ...,
  retstart = 0L,
  retmax = NA,
  retmode = "xml",
  rettype = "uilist",
  usehistory = is.null(retmax) || is.na(retmax),
  WebEnv = NULL,
  query_key = NULL,
  .cookies = NA,
  .paginate = 10000L,
  .path = NULL,
  .call = rlang::current_env()
)
}
\arguments{
\item{term}{search query.}

\item{db}{Entrez database name.}

\item{...}{additional API parameters (refer to Entrez documentation).
Any set to \code{NULL} are removed.}

\item{retstart}{integer: index of first result (starts from 0).
Ignored when \code{usehistory} is \code{TRUE}.}

\item{retmax}{integer: maximum number of results to return.
When \code{NA} this returns all results.
When \code{NULL}, uses the Entrez default (typically 20).
Note that it is possible that slightly more than \code{retmax} results will be returned
when paginating.
Ignored when \code{usehistory} is \code{TRUE}.}

\item{retmode}{character: currently only \code{"xml"} is supported.}

\item{rettype}{character: currently only \code{"uilist"} is supported.}

\item{usehistory}{logical: when \code{TRUE} use the history server to return the result.}

\item{WebEnv, query_key}{either characters to pass on as-is, or \code{\link{entrez_web_history}} objects.}

\item{.cookies}{path to persist cookies.
If \code{NULL}, cookies are not added to the request.
For helper functions: when \code{NA}, a temporary file is created (in this case only,
the temporary file will be cleaned up once all requests are performed).}

\item{.paginate}{controls how multiple API requests are used to complete the call.
Pagination is performed using the \code{retstart} and \code{retmax} API parameters.
When set to an integer, no more than \code{.pagniate} items will be requested per API call.
When \code{FALSE} or \code{0}, only one API request is sent.
Ignored when \code{usehistory} is \code{TRUE}.}

\item{.path}{path specification for saving raw responses.
See \code{path} argument of \code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}}.}

\item{.call}{call environment to use in error messages/traces.
See \link[rlang:topic-error-call]{rlang::topic-error-call} and the \code{call} argument of \code{\link[cli:cli_abort]{cli::cli_abort()}}.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}
}
\value{
id set object (either a web history token or explicit list).
}
\description{
The search term field names are documented in the EInfo API endpoint:
see \code{\link[=einfo]{einfo()}}.
}
\seealso{
\url{https://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.ESearch}

Other API methods: 
\code{\link{efetch}()},
\code{\link{einfo}()},
\code{\link{elink}()},
\code{\link{entrez_count}()},
\code{\link{entrez_translate}()},
\code{\link{epost}()},
\code{\link{esummary}()}
}
\concept{API methods}
