% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{esummary}
\alias{esummary}
\title{Fetch document summaries from Entrez}
\usage{
esummary(
  id_set,
  ...,
  retstart = 0L,
  retmax = NA,
  retmode = "xml",
  version = "2.0",
  .method = NA,
  .cookies = NA,
  .paginate = 5000L,
  .process = "identity",
  .path = NULL,
  .call = rlang::current_env()
)
}
\arguments{
\item{id_set}{ID set object.}

\item{...}{additional API parameters (refer to Entrez documentation).
Any set to \code{NULL} are removed.}

\item{retstart}{integer: index of first result (starts from 0).}

\item{retmax}{integer: maximum number of results to return.
When \code{NA} this returns all results. When \code{NULL}, uses the Entrez default (typically 20).
Note that when using pagination with web history, it is possible that slightly more than
\code{retmax} results will be returned.}

\item{retmode}{character: requested document format.}

\item{version}{character: requested format version.}

\item{.method}{HTTP verb. If \code{NA}, a sensible default is chosen based on the request parameters.}

\item{.cookies}{path to persist cookies.
If \code{NULL}, cookies are not added to the request.
For helper functions: when \code{NA}, a temporary file is created (in this case only,
the temporary file will be cleaned up once all requests are performed).}

\item{.paginate}{controls how multiple API requests are used to complete the call.
Pagination is performed using the \code{retstart} and \code{retmax} API parameters.
When set to an integer, no more than \code{.pagniate} items will be requested per API call.
When \code{FALSE} or \code{0}, only one API request is sent.}

\item{.process}{function that processes the API results.
Can be a function or builtin processor as described in \code{\link{process}}.}

\item{.path}{path specification for saving raw responses.
See \code{path} argument of \code{\link[httr2:req_perform_iterative]{httr2::req_perform_iterative()}}.}

\item{.call}{call environment to use in error messages/traces.
You only need to specify this in internal helper functions that don't need to be
mentioned in error messages.}
}
\description{
ESummary is faster than EFetch because it only interacts with the frontend
rather than the full database. It contains more limited information.
Consider adding \code{version = "2.0"} to request the revised output format.
}
\seealso{
Other API methods: 
\code{\link{efetch}()},
\code{\link{einfo}()},
\code{\link{elink}()},
\code{\link{entrez_count}()},
\code{\link{entrez_translate}()},
\code{\link{epost}()},
\code{\link{esearch}()}
}
\concept{API methods}
